add web server

+++ source/internal/echo/response.go
@@ -11,7 +11,7 @@ import (
 type ArrayResponse struct {
 	Code  int    `json:"code"`
 	Msg   string `json:"msg,omitempty"`
-	Total int    `json:"total,omitempty"`
+	Total int64  `json:"total,omitempty"`
 	Data  any    `json:"data,omitempty"`
 }
 
@@ -27,7 +27,7 @@ func WrapError(err1 error, err2 error) error {
 	return errors.New(err2.Error() + ": " + err1.Error())
 }
 
-func response(c echo.Context, code int, errCode int, err error, msg string, data any, total int) error {
+func response(c echo.Context, code int, errCode int, err error, msg string, data any, total int64) error {
 	c.Response().Header().Add("X-Request-Id", fmt.Sprintf("%v", c.Get("requestID")))
 	c.Set("code", errCode)
 	return WrapError(err, c.JSON(code, &ArrayResponse{Code: errCode, Msg: msg, Data: data, Total: total}))
